+++ package.json
@@ -18,6 +18,7 @@
 	"dependencies": {
 		"@ant-design/icons": "6.0.0",
 		"@ant-design/v5-patch-for-react-19": "1.0.3",
+		"@vanilla-extract/css": "^1.17.4",
 		"antd": "5.27.1",
 		"react": "^19.1.1",
 		"react-dom": "^19.1.1",
@@ -34,6 +35,7 @@
 		"@types/node": "^24.3.0",
 		"@types/react": "^19.1.10",
 		"@types/react-dom": "^19.1.7",
+		"@vanilla-extract/vite-plugin": "^5.1.1",
 		"@vitejs/plugin-react": "^5.0.0",
 		"eslint": "^9.33.0",
 		"eslint-config-prettier": "10.1.8",
